Refactor tests for configuration management: streamline ServerSettings and SearchAPISettings validation tests
